+++ chocolatey-visualstudio.extension/extensions/Install-VSChocolateyPackage.ps1
@@ -17,35 +17,45 @@ function Install-VSChocolateyPackage
         [string] $checksum64 = '',
         [string] $checksumType64 = '',
         [string] $logFilePath,
-        [switch] $assumeNewVS2017Installer
+        [switch] $assumeNewVS2017Installer,
+        [string] $installerFilePath
     )
 
-    Write-Debug "Running 'Install-VSChocolateyPackage' for $packageName with url:'$url', args:'$silentArgs', url64bit:'$url64bit', checksum:'$checksum', checksumType:'$checksumType', checksum64:'$checksum64', checksumType64:'$checksumType64', logFilePath:'$logFilePath'";
+    Write-Debug "Running 'Install-VSChocolateyPackage' for $packageName with url:'$url', args:'$silentArgs', url64bit:'$url64bit', checksum:'$checksum', checksumType:'$checksumType', checksum64:'$checksum64', checksumType64:'$checksumType64', logFilePath:'$logFilePath', installerFilePath:'$installerFilePath'";
 
-    $chocTempDir = $env:TEMP
-    $tempDir = Join-Path $chocTempDir "$packageName"
-    if ($env:packageVersion -ne $null) { $tempDir = Join-Path $tempDir "$env:packageVersion" }
+    if ($installerFilePath -eq '') {
+        $chocTempDir = $env:TEMP
+        $tempDir = Join-Path $chocTempDir "$packageName"
+        if ($env:packageVersion -ne $null) { $tempDir = Join-Path $tempDir "$env:packageVersion" }
 
-    if (![System.IO.Directory]::Exists($tempDir)) { [System.IO.Directory]::CreateDirectory($tempDir) | Out-Null }
-    $urlForFileNameDetermination = $url
-    if ($urlForFileNameDetermination -eq '') { $urlForFileNameDetermination = $url64bit }
-    if ($urlForFileNameDetermination -like '*.exe') { $localFileName = $urlForFileNameDetermination.Substring($urlForFileNameDetermination.LastIndexOfAny(@('/', '\')) + 1) }
-    else { $localFileName = 'vs_setup.exe' }
-    $localFilePath = Join-Path $tempDir $localFileName
+        if (![System.IO.Directory]::Exists($tempDir)) { [System.IO.Directory]::CreateDirectory($tempDir) | Out-Null }
+        $urlForFileNameDetermination = $url
+        if ($urlForFileNameDetermination -eq '') { $urlForFileNameDetermination = $url64bit }
+        if ($urlForFileNameDetermination -like '*.exe') { $localFileName = $urlForFileNameDetermination.Substring($urlForFileNameDetermination.LastIndexOfAny(@('/', '\')) + 1) }
+        else { $localFileName = 'vs_setup.exe' }
+        $localFilePath = Join-Path $tempDir $localFileName
 
-    $arguments = @{
-        packageName = $packageName
-        fileFullPath = $localFilePath
-        url = $url
-        url64bit = $url64bit
-        checksum = $checksum
-        checksumType = $checksumType
-        checksum64 = $checksum64
-        checksumType64 = $checksumType64
+        Write-Verbose "Downloading the installer executable"
+        $arguments = @{
+            packageName = $packageName
+            fileFullPath = $localFilePath
+            url = $url
+            url64bit = $url64bit
+            checksum = $checksum
+            checksumType = $checksumType
+            checksum64 = $checksum64
+            checksumType64 = $checksumType64
+        }
+        Set-StrictMode -Off
+        Get-ChocolateyWebFile @arguments | Out-Null
+        Set-StrictMode -Version 2
+    } else {
+        if (-not (Test-Path -Path $installerFilePath)) {
+            throw "The local installer executable does not exist: $installerFilePath"
+        }
+        Write-Verbose "Using a local installer executable: $installerFilePath"
+        $localFilePath = $installerFilePath
     }
-    Set-StrictMode -Off
-    Get-ChocolateyWebFile @arguments | Out-Null
-    Set-StrictMode -Version 2
 
     $arguments = @{
         packageName = $packageName