+++ chocolatey-visualstudio.extension/CHANGELOG.md
@@ -1,5 +1,9 @@
 ﻿# CHANGELOG
 
+## Version 1.5.1
+- Changed the method of locating the VS 2017 installer during modify and uninstall operations to not depend on Uninstall registry
+  keys anymore. This avoids the problem caused by registry key changes in a recent VS 2017 update.
+
 ## Version 1.5.0
 - New helpers: Add-VisualStudioComponent, Remove-VisualStudioComponent
 - New package parameter: '--layout D:\Path' can be used to create an offline installation source ("layout").